Examples for using Docker+Caddy to host site
Provide a working `Dockerfile` and an _example_ `Caddyfile` to help
people bring up a site quickly and fairly easily.

This is entirely optional and not used by me in Production deployments
at this time.

The resulting image brings up a Caddy webserver with all content present
and Caddy automatically fetches the Let's Encrypt certificates for the
hostname.  With some redirects on the local router, I can test this as
fully working on my laptop.

Describe how to use this in a real deployment, and point to how to
improve things further.

Caddy is a bit fiddly on hostname matching when you vary port numbers
under it, which is why I leave it using 80/443 inside the Container and
so running as contained-root, instead of completely unprivileged.
Provide instructions for a second example on configuring Caddy to
respond to any hostname and not handle certs, for deploying behind
TLS-terminating load-balancers.