Make pcregrep an optional dependency
Check for pcregrep and use it if present; else use python3 short in-line
script.  The pcregrep line at least is easier for non-Python programmers
to read.

The bin/update-sites script is still marginally better off in bash than
in Python, but not by much, and this can stand to be replaced if it gets
any more complex.