Tooling to create Exim-Maintainers-Keyring.asc
This is the workflow I previously communicated to exim maintainers and
distributors, adapted a little to also use the keys in this repo.

It's not a perfect match for openpgpkey-control, but it is a useful
indicator of how to put together a group keyring: we want
cross-signatures to be intact in the published ring, but we don't want
spammy or abusive signatures distributed.

So we export the keys which we care about, put together a temporary
keyring, and then from that perform a minimizing export, so that all
cross-signatures are present.