Bug-fix multiple keys; support multi keys/domains
To make it easier to use globs to import all the .asc files in a
directory, switch the parsing of `--keys-file` to take one or more
parameters.  Then do the same for `--domain`, so that a site which
serves multiple domains with the same content can have all the PGP keys
in one place.

Bug-fix for creating a sorted list of email addresses: I was missing the
dunder method to support less-than comparison, needed for Python
sorting of custom classes.  Added `__lt__` method.